refactor: move `templates` to `commands`
