+++ src/main/java/pw/mihou/amelia/commands/RemoveCommand.kt
@@ -5,10 +5,10 @@ import org.javacord.api.entity.message.embed.EmbedBuilder
 import org.javacord.api.interaction.SlashCommandOption
 import org.javacord.api.interaction.SlashCommandOptionType
 import pw.mihou.amelia.commands.middlewares.Middlewares
+import pw.mihou.amelia.commands.templates.TemplateMessages
 import pw.mihou.amelia.db.FeedDatabase
 import pw.mihou.amelia.db.methods.feeds.Feeds
 import pw.mihou.amelia.db.models.FeedModel
-import pw.mihou.amelia.templates.TemplateMessages
 import pw.mihou.amelia.utility.confirmationMenu
 import pw.mihou.amelia.utility.redactListLink
 import pw.mihou.nexus.features.command.facade.NexusCommandEvent
