+++ src/main/java/pw/mihou/amelia/commands/components/LoadingComponent.kt
@@ -1,6 +1,6 @@
 package pw.mihou.amelia.commands.components
 
-import pw.mihou.amelia.templates.TemplateMessages
+import pw.mihou.amelia.commands.templates.TemplateMessages
 import pw.mihou.reakt.Reakt
 
 val Reakt.Document.Loading get() =
