+++ src/main/java/pw/mihou/amelia/commands/subcommands/RegisterAuthorSubcommand.kt
@@ -9,8 +9,8 @@ import pw.mihou.amelia.commands.components.PlainEmbed
 import pw.mihou.amelia.commands.components.UserResultEmbed
 import pw.mihou.amelia.commands.hooks.useDeletable
 import pw.mihou.amelia.commands.hooks.useLoader
+import pw.mihou.amelia.commands.templates.TemplateMessages
 import pw.mihou.amelia.db.methods.feeds.Feeds
-import pw.mihou.amelia.templates.TemplateMessages
 import pw.mihou.models.user.UserResultOrAuthor
 import pw.mihou.nexus.coroutines.utils.coroutine
 import pw.mihou.nexus.features.command.facade.NexusCommandEvent
