+++ src/main/java/pw/mihou/amelia/commands/subcommands/RegisterListSubcommand.kt
@@ -2,12 +2,12 @@ package pw.mihou.amelia.commands.subcommands
 
 import org.javacord.api.interaction.SlashCommandInteractionOption
 import org.javacord.api.util.logging.ExceptionLogger
+import pw.mihou.amelia.commands.templates.TemplateMessages
 import pw.mihou.amelia.db.FeedDatabase
 import pw.mihou.amelia.db.models.FeedModel
 import pw.mihou.amelia.extensions.params
 import pw.mihou.amelia.extensions.toUrl
 import pw.mihou.amelia.rss.reader.RssReader
-import pw.mihou.amelia.templates.TemplateMessages
 import pw.mihou.nexus.features.command.facade.NexusCommandEvent
 
 object RegisterListSubcommand {
